islet_rmm::exception

Module trap

Modules

syndrome

Structs

Info

Enums

Kind

Source

Constants

RET_TO_REC

RET_TO_RMM

Functions

handle_exception

This function is called when an exception occurs from CurrentSPEL0, CurrentSPELx. The info parameter specifies source (first 16 bits) and kind (following 16 bits) of the exception. The esr has the value of a syndrome register (ESR_ELx) holding the cause of the Synchronous and SError exception. The tf has the TrapFrame of current context.

handle_lower_exception

This function is called when an exception occurs from LowerAArch64. To enter RMM (EL2), return 1. Otherwise, return 0 to go back to EL1. The info parameter specifies source (first 16 bits) and kind (following 16 bits) of the exception. The esr has the value of a syndrome register (ESR_ELx) holding the cause of the Synchronous and SError exception. The rec has the Rec context. The tf has the TrapFrame of current context.