islet_rmm

Function rmm_exit

pub unsafe fn rmm_exit(args: [usize; 4]) -> [usize; 4]

Call rmm_exit within exception/vectors.s and jumps to EL1.

Currently, this function gets [0usize; 3] as an argument to initialize x0, x1 and x2 registers.

When an exception occurs and the flow comes back to EL2 through rmm_enter, x0, x1 and x2 registers might be changed to contain additional information set from handle_lower_exception. These are the return values of this function. The return value encodes: [rmi::RET_XXX, ret_val1, ret_val2] In most cases, the function returns [rmi::RET_SUCCESS, _, _] pagefault returns [rmi::RET_PAGE_FAULT, faulted address, _]

Safety

This function alters the processor’s execution level by jumping to EL1; the caller must ensure that the system is in a correct state for this transition.