Trait islet_rmm::host::Accessor

pub trait Accessor {
    // Provided methods
    fn acquire(ptr: usize) -> bool { ... }
    fn release(ptr: usize) -> bool { ... }
    fn validate(&self) -> bool { ... }
}

This trait is used to enforce security checks for physical region allocated by the host. This is used for PointerGuard which is not able to modify data.

Provided Methods

fn acquire(ptr: usize) -> bool

Try to do page-relevant stuff (e.g., RMM map). returns true only if everything goes well.

fn release(ptr: usize) -> bool

Try to clean up page-relevant stuff done by acquire. Structs that implement this trait must synchronize this function with acquire. returns true only if everything goes well.

fn validate(&self) -> bool

Validate each field in a struct that implements this trait. returns true only if everything goes well.