Macro islet_rmm::get_granule

macro_rules! get_granule {
    ($addr:expr) => { ... };
}

Safety / Usage: “granule transaction” a set of APIs that define how to access “granule” and contents inside it.

Goal

The high-level goal of these APIs is to enforce that read/write from/to “granule” is only allowed after getting a proper granule lock.

Usage: take the example of REALM_DESTROY. (1) let mut g = get_granule_if!(addr, GranuleStatus::RD); –> getting a granule while holding a lock. (precisely, it gets EntryGuardentry::Inner) (2) … do something under g … (3) set_granule(&mut g, GranuleStatus::Delegated) –> do a state transition when everything goes ok.

this can be seen as a typical transaction that defines a way to access granule. (granule transaction)

to open a transaction, you can use either get_granule or get_granule_if or set_state_and_get_granule.

to modify a granule and close a transaction, set_granule can be used.

Safety:

In the above usage, set_granule() takes &entry::Inner as input, which can only be acquired by get_granule_if!. This means that you can access granule only after get_granule_if!, and this is ensured by this API design.

TODO(Optional): currently, handling a granule transaction is in charge of developers. e.g., (1) they should invoke set_granule() at the right moment, (2) EntryGuard still lives after set_granule(). we might be able to free them from it via Rust features.

TODO(Must): currently, these APIs do not involve in mapping some address into RMM page table. callers must map some physical address, if needed, before/after using these APIs. we need to define and use a secure interface to map some address and do set_granule().

Note:

why is it using macros, not functions? writing a function that returns EntryGuard<> requires cloning EntryGuard. To get around this, macros are currently used. we might be able to figure out a better way to define get_granule_*! macros. get_granule!(addr: a physical address)

when success, returns EntryGuard<entry::Inner> allowing an access to “Granule”.