islet_rmm::rec::context

Struct Context

#[repr(C)]
pub struct Context {
    pub gp_regs: [u64; 31],
    pub elr_el2: u64,
    pub spsr_el2: u64,
    pub sys_regs: SystemRegister,
    pub mdcr_el2: u64,
    pub gic_state: GICRegister,
    pub timer: TimerRegister,
    pub simd: SimdContext,
    pub pauth: PauthRegister,
    pub pmu: PmuRegister,
}

Fields
gp_regs: [u64; 31]elr_el2: u64spsr_el2: u64sys_regs: SystemRegistermdcr_el2: u64gic_state: GICRegistertimer: TimerRegistersimd: SimdContextpauth: PauthRegisterpmu: PmuRegister
Implementations

impl Context

pub fn new() -> Self

pub unsafe fn into_current(rec: &Rec<'_>)

Restores the current execution context from the given Rec.

Safety

This function modifies processor-specific registers and state; ensure that this is safe in the current execution context.

pub unsafe fn from_current(rec: &mut Rec<'_>)

Saves the current execution context into the given Rec record.

Safety

This function reads and modifies processor-specific registers and state; ensure that this is appropriate in the current execution context.