Module islet_rmm::rmi::rec::run

Structs

Run

The structure holds data passsed between the Host and the RMM on Realm Execution Context (REC) entry and exit.

Constants

NR_GPRS

REC_ENTRY_FLAG_EMUL_MMIO

Whether the host has completed emulation for an Emulatable Data Abort. val 0: Host has not completed emulation for an Emulatable Abort. val 1: Host has completed emulation for an Emulatable Abort.

REC_ENTRY_FLAG_INJECT_SEA

Whether to inject a Synchronous External Abort (SEA) into the Realm. val 0: Do not inject an SEA into the Realm. val 1: Inject an SEA into the Realm.

REC_ENTRY_FLAG_RIPAS_RESPONSE

Host response to RIPAS change request val 0: ACCEPT val 1: REJECT

REC_ENTRY_FLAG_TRAP_WFE

Whether to trap WFE execution by the Realm. val 0: Trap is disabled. val 1: Trap is enabled.

REC_ENTRY_FLAG_TRAP_WFI

Whether to trap WFI execution by the Realm. val 0: Trap is disabled. val 1: Trap is enabled.