Function islet_rmm::exception::trap::handle_lower_exception

#[no_mangle]
pub extern "C" fn handle_lower_exception(
    info: Info,
    esr: u32,
    vcpu: &mut VCPU<Context>,
    tf: &mut TrapFrame
) -> u64

This function is called when an exception occurs from LowerAArch64. To enter RMM (EL2), return 1. Otherwise, return 0 to go back to EL1. The info parameter specifies source (first 16 bits) and kind (following 16 bits) of the exception. The esr has the value of a syndrome register (ESR_ELx) holding the cause of the Synchronous and SError exception. The vcpu has the VCPU context. The tf has the TrapFrame of current context.

Do not write sys_regs of VCPU here. (ref. HANDLE_LOWER in vectors.s)